Move delete stuff back to AddStepPage to facilitate working with history
Minor changes

Rm i18n to res conflicts
